Refactor Lexer, Parser & TokenKind
This commit includes several changes to improve the code of the Lexer,
Parser, and TokenKind classes.

Firstly, we have consolidated the FLOAT and INTEGER TokenKinds into a
single NUMBER group, as the separation was not necessary.

Secondly, we have implemented the Iterator interface in the Lexer class,
which will make it easier to iterate over the tokens produced by the
lexer.

Lastly, we have improved the parseFactor method in the Parser class by
replacing the if-else statements with a switch statement, which is more
concise and easier to read.